requirements.txt
# 基础依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
einops>=0.8.0
Pillow>=9.0.0
opencv-python>=4.7.0
tqdm>=4.65.0

# 配置和工具
hydra-core>=1.3.0
omegaconf>=2.3.0
pyrootutils>=1.0.4
colorlog>=6.7.0

# 模型相关
transformers>=4.30.0
groundingdino-py>=0.4.0
ultralytics>=8.0.0
segment-anything>=1.0
sam2>=0.1.0
sahi>=0.1.0

# 文件处理
python-dotenv>=1.0.0
aria2p>=0.11.0 